Client change game mode event now fires for all players
